refactor: move Server to its own file
